+++ src/util/solver.cpp
@@ -34,6 +34,8 @@ double sparse_llt_solver::trace(const SparseMatrix<double, 0, int> &M,
         QU = R_supernodal.solve(U);
       else if (solver_type == 3)
         QU = R_cholmod_ldlt.solve(U);
+      else if (solver_type == 8)
+        QU = R_sparseqr.solve(U);
 #ifdef __APPLE__
       else if (solver_type == 4)
         QU = R_accelerate.solve(U);
@@ -99,6 +101,8 @@ double sparse_llt_solver::trace2(const SparseMatrix<double, 0, int> &A,
         QU = R_supernodal.solve(U);
       else if (solver_type == 3)
         QU = R_cholmod_ldlt.solve(U);
+      else if (solver_type == 8)
+        QU = R_sparseqr.solve(U);
 #ifdef __APPLE__
       else if (solver_type == 4)
         QU = R_accelerate.solve(U);